requirements.txt
# Dependensi aplikasi Python
# Tambahkan package yang Anda butuhkan di bawah ini

# Web Framework
flask==2.3.3

# Database
pymongo==4.13.2
neo4j==5.28.1
python-dotenv==1.1.1

# Contoh dependensi umum:
# requests==2.31.0
# pandas==2.0.3
# numpy==1.24.3
# matplotlib==3.7.2
# tkinter  # Biasanya sudah terinstall dengan Python 